# 56 methods too slow