--- a/ext_list/dict_operations.py
+++ b/ext_list/dict_operations.py
@@ -141,7 +141,7 @@ class _DictOperation(List[T]):  # type: ignore
         return result
 
     def rename_keys(self, rename_keys: dict[Hashable, Hashable]) -> Iterable[T]:
-        def __copy_objects() -> Iterable[Any]:
+        def __copy_object() -> Iterable[Any]:
             if isinstance(self[0], dict):
                 return [dict(element) for element in self]  # type: ignore[assignment]
 
@@ -164,12 +164,12 @@ class _DictOperation(List[T]):  # type: ignore
         if not base.is_indexable(self):
             raise TypeError
 
-        copied_elements = __copy_objects()
+        copied_elements = __copy_object()
 
         return self.__class__([__swap_keys(element, rename_keys) for element in copied_elements])
 
     def map_for_keys(self, keys: list[Hashable], function: Callable[[Any], Any] | type, *args: Any) -> Iterable[dict[Any, Any]]:
-        def __copy_objects() -> Iterable[Any]:
+        def __copy_object() -> Iterable[Any]:
             if isinstance(self[0], dict):
                 return [dict(element) for element in self]  # type: ignore[assignment]
 
@@ -187,7 +187,7 @@ class _DictOperation(List[T]):  # type: ignore
 
         result: Iterable[dict[Any, Any]] = self.__class__()
 
-        for element in __copy_objects():
+        for element in __copy_object():
             for key in keys:
                 element[key] = function(element[key], *args)  # type: ignore[attr-defined]